+++ CHANGELOG.md
@@ -7,6 +7,9 @@ First release.
 > This section will be removed after the beta phase. <br>
 > Note that semantic versioning rules are not strictly followed during this phase.
 
+- feature/next-match: Filter: Add support for prev/next-match
+- feature/next-match: Filter: New mode 'mark' (like 'dim' but does not gray out)
+
 - v0.12.1: Fix flat source format for positional args.
 
 - v0.12.0: Add `deep`, `resetLazy`, and `collapseOthers` options to